Getting rid of duplicate logic for extracting values for parameters
