+++ Source/JavaScript/Applications/queries/IObservableQueryFor.ts
@@ -4,7 +4,7 @@
 import { ObservableQuerySubscription } from './ObservableQuerySubscription';
 import { QueryResult } from './QueryResult';
 import { IQuery } from './IQuery';
-import { ParameterDescriptor } from '../reflection/ParameterDescriptor';
+import { IHaveParameters } from '../reflection/IHaveParameters';
 
 /**
  * The delegate type representing the callback of result from the server.
@@ -16,10 +16,9 @@ export type OnNextResult<TDataType> = (data: TDataType) => void;
  * @template TDataType Type of model the query is for.
  * @template TArguments Optional type of arguments to use for the query.
  */
-export interface IObservableQueryFor<TDataType, TArguments = object> extends IQuery {
+export interface IObservableQueryFor<TDataType, TArguments = object> extends IQuery, IHaveParameters {
     readonly route: string;
     readonly requiredRequestParameters: string[];
-    readonly parameterDescriptors: ParameterDescriptor[];
     readonly defaultValue: TDataType;
 
     /**
