+++ Source/JavaScript/Applications/queries/IQueryFor.ts
@@ -3,17 +3,16 @@
 
 import { QueryResult } from './QueryResult';
 import { IQuery } from './IQuery';
-import { ParameterDescriptor } from '../reflection/ParameterDescriptor';
+import { IHaveParameters } from '../reflection/IHaveParameters';
 
 /**
  * Defines the base of a query.
  * @template TDataType Type of model the query is for.
  * @template TArguments Optional type of arguments to use for the query.
  */
-export interface IQueryFor<TDataType, TArguments = object> extends IQuery {
+export interface IQueryFor<TDataType, TArguments = object> extends IQuery, IHaveParameters {
     readonly route: string;
     readonly requiredRequestParameters: string[];
-    readonly parameterDescriptors: ParameterDescriptor[];
     readonly defaultValue: TDataType;
 
     /**
