+++ Source/JavaScript/Applications/queries/ObservableQueryFor.ts
@@ -18,6 +18,7 @@ import { joinPaths } from '../joinPaths';
 import { UrlHelpers } from '../UrlHelpers';
 import { GetHttpHeaders } from '../GetHttpHeaders';
 import { ParameterDescriptor } from '../reflection/ParameterDescriptor';
+import { ParametersHelper } from '../reflection/ParametersHelper';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
@@ -134,7 +135,7 @@ export abstract class ObservableQueryFor<TDataType, TParameters = object> implem
         }
 
         // Collect parameter values from parameterDescriptors that are set
-        const parameterValues = this.collectParameterValues();
+        const parameterValues = ParametersHelper.collectParameterValues(this);
 
         const queryParams = UrlHelpers.buildQueryParams({ ...unusedParameters, ...parameterValues }, additionalParams);
         const queryString = queryParams.toString();
@@ -225,16 +226,5 @@ export abstract class ObservableQueryFor<TDataType, TParameters = object> implem
         } else {
             result.data = JsonSerializer.deserializeFromInstance(this.modelType, result.data);
         }
-    }
-
-    private collectParameterValues(): Record<string, unknown> {
-        const values: Record<string, unknown> = {};
-        for (const descriptor of this.parameterDescriptors) {
-            const value = this[descriptor.name];
-            if (value !== undefined && value !== null) {
-                values[descriptor.name] = value;
-            }
-        }
-        return values;
     }
 }
