+++ Source/JavaScript/Applications/reflection/index.ts
@@ -1,5 +1,7 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+export * from './IHaveParameters';
 export * from './ParameterDescriptor';
+export * from './ParametersHelper';
 export * from './PropertyDescriptor';
